add util function for retry with backoff
