+++ packages/node-core/src/utils/promise.ts
@@ -28,3 +28,31 @@ export async function timeout<T>(promise: Promise<T>, sec: number): Promise<T> {
     }),
   ]);
 }
+
+function backoff(attempt: number): number {
+  return Math.pow(2, attempt) * 1000; // Exponential backoff
+}
+
+export function retryWithBackoff<T>(
+  tryFunction: () => Promise<T>,
+  onError: (error: any) => void,
+  onMaxAttempts: () => void,
+  attempt = 0,
+  maxAttempts = 5
+): NodeJS.Timeout | undefined {
+  if (attempt >= maxAttempts) {
+    onMaxAttempts();
+  } else {
+    // eslint-disable-next-line @typescript-eslint/no-misused-promises
+    const timeout = setTimeout(async () => {
+      try {
+        await tryFunction();
+      } catch (error) {
+        onError(error);
+        retryWithBackoff(tryFunction, onError, onMaxAttempts, attempt + 1, maxAttempts);
+      }
+    }, backoff(attempt));
+
+    return timeout;
+  }
+}
