+++ packages/node/src/indexer/indexer.manager.spec.ts
@@ -16,7 +16,6 @@ import {
   StoreCacheService,
   ConnectionPoolStateManager,
   MmrQueryService,
-  RetryManager,
   IProjectUpgradeService,
 } from '@subql/node-core';
 import { Sequelize } from '@subql/x-sequelize';
@@ -167,7 +166,6 @@ function createIndexerManager(
     connectionPoolService,
     eventEmitter,
     nodeConfig,
-    new RetryManager(),
   );
   const dsProcessorService = new DsProcessorService(project, nodeConfig);
   const dynamicDsService = new DynamicDsService(dsProcessorService, project);
@@ -248,7 +246,6 @@ describe('IndexerManager', () => {
       new ConnectionPoolService<ApiPromiseConnection>(
         nodeConfig,
         new ConnectionPoolStateManager(),
-        new RetryManager(),
       ),
       nodeConfig,
     );
@@ -263,7 +260,6 @@ describe('IndexerManager', () => {
       new ConnectionPoolService<ApiPromiseConnection>(
         nodeConfig,
         new ConnectionPoolStateManager(),
-        new RetryManager(),
       ),
       nodeConfig,
     );
