+++ packages/node/src/indexer/indexer.module.ts
@@ -16,7 +16,6 @@ import {
   NodeConfig,
   PgMmrCacheService,
   MmrQueryService,
-  RetryManager,
 } from '@subql/node-core';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { ApiService } from './api.service';
@@ -36,7 +35,6 @@ import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlock
     IndexerManager,
     StoreCacheService,
     StoreService,
-    RetryManager,
     {
       provide: ConnectionPoolStateManager,
       useFactory: () => {
@@ -54,14 +52,12 @@ import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlock
         connectionPoolService: ConnectionPoolService<ApiPromiseConnection>,
         eventEmitter: EventEmitter2,
         nodeConfig: NodeConfig,
-        retryManager: RetryManager,
       ) => {
         const apiService = new ApiService(
           project,
           connectionPoolService,
           eventEmitter,
           nodeConfig,
-          retryManager,
         );
         await apiService.init();
         return apiService;
@@ -71,7 +67,6 @@ import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlock
         ConnectionPoolService,
         EventEmitter2,
         NodeConfig,
-        RetryManager,
       ],
     },
     SandboxService,
