+++ packages/node/src/subcommands/testing.module.ts
@@ -10,7 +10,6 @@ import {
   DbModule,
   NodeConfig,
   PoiService,
-  RetryManager,
   StoreService,
   TestRunner,
 } from '@subql/node-core';
@@ -39,7 +38,6 @@ import { MetaModule } from '../meta/meta.module';
     UnfinalizedBlocksService,
     ConnectionPoolStateManager,
     ConnectionPoolService,
-    RetryManager,
     {
       provide: 'IProjectService',
       useClass: ProjectService,
@@ -51,14 +49,12 @@ import { MetaModule } from '../meta/meta.module';
         connectionPoolService: ConnectionPoolService<ApiPromiseConnection>,
         eventEmitter: EventEmitter2,
         config: NodeConfig,
-        retryManager: RetryManager,
       ) => {
         const apiService = new ApiService(
           project,
           connectionPoolService,
           eventEmitter,
           config,
-          retryManager,
         );
         await apiService.init();
         return apiService;
@@ -68,7 +64,6 @@ import { MetaModule } from '../meta/meta.module';
         ConnectionPoolService,
         EventEmitter2,
         NodeConfig,
-        RetryManager,
       ],
     },
     SchedulerRegistry,
